cleaned up some code to be more in the style of norm

+++ inc/scene_elements.h
@@ -18,7 +18,7 @@ typedef struct s_camera
 	float	fov;
 	float	viewport_height;
 	float	viewport_width;
-	uint8_t	samples_per_pixel;
+	uint16_t	samples_per_pixel;
 	uint8_t	max_rays;
 }	t_camera;
 
@@ -68,7 +68,7 @@ typedef struct s_hit_record
 }	t_hit_record;
 
 t_vec3f		at(t_ray r, float t);
-t_vec3f		ray_color(const t_ray r, const t_hittables *htbl, uint8_t depth);
+t_vec3f		ray_color(const t_ray r, const t_hittables *htbl, uint16_t depth);
 void		render(const t_hittables *htbl, const t_camera cam, const t_image img);
 float		hit_sphere(const t_sphere s, const t_ray r);
 t_camera	init_camera(const t_image *img);
